+++ app/components/content/Docs/Card/DocsCardExample.vue
@@ -14,7 +14,7 @@
               <p class="text-muted-foreground">Send notifications to device.</p>
             </UiLabel>
             <div class="ml-auto">
-              <UiSwitch id="push" v-model:checked="push" />
+              <UiSwitch id="push" v-model="push" />
             </div>
           </div>