+++ app/components/content/Docs/Switch/DocsSwitchOrigin.vue
@@ -19,7 +19,7 @@
       <div
         class="inline-flex items-center gap-2 [--primary:var(--color-indigo-500)] [--ring:var(--color-indigo-300)] dark:[--primary:var(--color-indigo-500)] dark:[--ring:var(--color-indigo-900)]"
       >
-        <UiSwitch :id="id + 'colored-switch'" :default-checked="true" />
+        <UiSwitch :id="id + 'colored-switch'" :default-value="true" />
         <UiLabel :for="id + 'colored-switch'" class="sr-only">Colored switch</UiLabel>
       </div>